pyproject.toml
[tool.poetry]
name = "notifications-api"
version = "0.1.0"
description = "Notify.gov backend"
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12.2"
alembic = "==1.13.1"
amqp = "==5.2.0"
beautifulsoup4 = "==4.12.3"
boto3 = "^1.34.114"
botocore = "^1.34.115"
cachetools = "==5.3.3"
celery = {version = "==5.4.0", extras = ["redis"]}
certifi = ">=2022.12.7"
cffi = "==1.16.0"
charset-normalizer = "^3.1.0"
click = "==8.1.7"
click-datetime = "==0.2"
click-didyoumean = "==0.3.1"
click-plugins = "==1.1.1"
click-repl = "==0.3.0"
deprecated = "==1.2.14"
eventlet = "==0.36.1"
expiringdict = "==1.2.2"
flask = "~=3.0"
flask-bcrypt = "==1.0.1"
flask-marshmallow = "==1.2.1"
flask-migrate = "==4.0.7"
flask-redis = "==0.4.0"
flask-sqlalchemy = "==3.1.1"
gunicorn = {version = "==22.0.0", extras = ["eventlet"]}
iso8601 = "==2.1.0"
jsonschema = {version = "==4.22.0", extras = ["format"]}
lxml = "==5.2.2"
marshmallow = "==3.21.2"
marshmallow-sqlalchemy = "==1.0.0"
newrelic = "*"
notifications-python-client = "==9.1.0"
oscrypto = "==1.3.0"
packaging = "==24.0"
poetry-dotenv-plugin = "==0.2.0"
psycopg2-binary = "==2.9.9"
pyjwt = "==2.8.0"
python-dotenv = "==1.0.1"
sqlalchemy = "==2.0.30"
werkzeug = "^3.0.3"
faker = "^25.3.0"
async-timeout = "^4.0.3"
bleach = "^6.1.0"
geojson = "^3.1.0"
govuk-bank-holidays = "^0.14"
numpy = "^1.26.4"
ordered-set = "^4.1.0"
phonenumbers = "^8.13.36"
python-json-logger = "^2.0.7"
pytz = "^2024.1"
regex = "^2024.5.15"
shapely = "^2.0.4"
smartypants = "^2.0.1"
mistune = "0.8.4"
blinker = "^1.8.2"
cryptography = "^42.0.7"
idna = "^3.7"
jmespath = "^1.0.1"
markupsafe = "^2.1.5"
pycparser = "^2.22"
python-dateutil = "^2.9.0.post0"
pyyaml = "^6.0.1"
s3transfer = "^0.10.1"
six = "^1.16.0"
urllib3 = "^2.2.1"
webencodings = "^0.5.1"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.4"
redis = "^5.0.4"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
awscli = "^1.29.74"
bandit = "*"
black = "^24.3.0"
cloudfoundry-client = "*"
exceptiongroup = "==1.2.1"
flake8 = "^7.0.0"
flake8-bugbear = "^24.1.17"
freezegun = "^1.5.1"
honcho = "*"
isort = "^5.13.2"
jinja2-cli = {version = "==0.8.2", extras = ["yaml"]}
moto = "==5.0.8"
pip-audit = "*"
pre-commit = "^3.7.1"
pytest = "^8.2.1"
pytest-env = "^1.1.3"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
pytest-xdist = "^3.5.0"
radon = "^6.0.1"
requests-mock = "^1.11.0"
setuptools = "^70.0.0"
sqlalchemy-utils = "^0.41.2"
vulture = "^2.10"
detect-secrets = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
